updated to 1.52

+++ SDK/install.sh
@@ -3,26 +3,32 @@
 # Install prerequisites if root
 if [ "$EUID" -ne 0 ]; then
   echo "Not root, skipping update and install"
+  exit
 else
   apt-get update
   apt-get -y install binutils gcc make
 fi
 
 # Delete directory if it exists and make empty directory
-rm -rf /opt/ps4sdk || true
-mkdir -p /opt/ps4sdk
+if [ "$PWD" != "/opt/ps4sdk" ]; then
+  rm -rf /opt/ps4sdk || true
+  mkdir -p /opt/ps4sdk
+fi
 
 # Build SDK
-pushd libPS4
+cd libPS4 || (echo "Unable to enter subdirectory" && exit)
 make
-popd
+cd ..
 
 # Copy compiled SDK
-cp -r libPS4 /opt/ps4sdk
+if [ "$PWD" != "/opt/ps4sdk" ]; then
+  cp -r libPS4 /opt/ps4sdk
+  cp install.sh /opt/ps4sdk/update.sh
+fi
 
 # Clear path from .bashrc if it's there
-sed -i "s/^\s*export PS4SDK=.*//gm" ~/.bashrc
+sed -i "s/^\s*export PS4SDK=.*//gm" /etc/profile
 
 # Add to paths
-echo "export PS4SDK=/opt/ps4sdk" | tee -a ~/.bashrc
+echo "export PS4SDK=/opt/ps4sdk" | tee -a /etc/profile
 export PS4SDK=/opt/ps4sdk
